requirements.txt
openai~=1.14.0
tiktoken~=0.6.0
numpy~=1.26.4
python_dotenv~=1.0.1
pymongo~=4.6.2
Flask~=3.0.2